test (일관성을 위해 HttpRequestParser에서 쿠키도 처리하도록 테스트 추가 및 변경!)
